+++ src/main/java/Table/ReservationPayment.java
@@ -2,18 +2,18 @@ package Table;
 
 public class ReservationPayment extends Reservation {
     private int amountCustomer;
-    private int total_price;
+    private Double total_price;
     private String depositStatus;
     private String arrearsStatus;
 
     public ReservationPayment(){
         this.amountCustomer = 0;
-        this.total_price = 0 ;
+        this.total_price = 0.0 ;
         this.depositStatus = "Not paid";
         this.arrearsStatus = "Not paid";
     }
 
-    public ReservationPayment(String reservationCode, String tourID, String tourName, String customerID, String customerName, String employeeID, String employeeName, int amountCustomer, int total_price, String depositStatus, String arrearsStatus) {
+    public ReservationPayment(String reservationCode, String tourID, String tourName, String customerID, String customerName, String employeeID, String employeeName, int amountCustomer, Double total_price, String depositStatus, String arrearsStatus) {
         super(reservationCode, tourID, tourName, customerID, customerName, employeeID, employeeName);
         this.amountCustomer = amountCustomer;
         this.total_price = total_price;
@@ -45,15 +45,15 @@ public class ReservationPayment extends Reservation {
         this.arrearsStatus = arrearsStatus;
     }
 
-    public int getTotal_price() {
+    public Double getTotal_price() {
         return total_price;
     }
 
-    public void setTotal_price(int total_price) {
+    public void setTotal_price(Double total_price) {
         this.total_price = total_price;
     }
 
-    public int calculateTotalPrice(int tourPrice){
+    public Double calculateTotalPrice(Double tourPrice){
         return amountCustomer * tourPrice;
     }
 }
